+++ app/components/mobile/ge-sorting-dialog-content.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry/components/mobile/ge-sorting-dialog-content';