+++ tests/dummy/app/controllers/log-service-examples/settings-example.js
@@ -56,13 +56,6 @@ export default ApplicationLogListFormController.extend({
 
   init() {
     this._super(...arguments);
-    this.get('logService').on('error', this, this._refreshList);
-    this.get('logService').on('warn', this, this._refreshList);
-    this.get('logService').on('info', this, this._refreshList);
-    this.get('logService').on('log', this, this._refreshList);
-    this.get('logService').on('debug', this, this._refreshList);
-    this.get('logService').on('deprecation', this, this._refreshList);
-    this.get('logService').on('promise', this, this._refreshList);
 
     let i18nService = this.get('i18n');
     let messagesToRefreshList = [
@@ -78,16 +71,6 @@ export default ApplicationLogListFormController.extend({
     ];
 
     this.set('messagesToRefreshList', messagesToRefreshList);
-  },
-
-  willDestroy() {
-    this.get('logService').off('error', this, this._refreshList);
-    this.get('logService').off('warn', this, this._refreshList);
-    this.get('logService').off('info', this, this._refreshList);
-    this.get('logService').off('log', this, this._refreshList);
-    this.get('logService').off('debug', this, this._refreshList);
-    this.get('logService').off('deprecation', this, this._refreshList);
-    this.get('logService').off('promise', this, this._refreshList);
   },
 
   actions: {
@@ -250,29 +233,5 @@ export default ApplicationLogListFormController.extend({
   */
   _generateUniqueMessagePrefix() {
     return '№' + Ember.generateGuid(null, '') + ': ';
-  },
-
-  /**
-    Refreshes list of log messages on form.
-
-    @method _refreshList
-    @private
-  */
-  _refreshList(applicationLogModel) {
-    let messagesToRefreshList = this.get('messagesToRefreshList');
-    if (applicationLogModel) {
-      let message = applicationLogModel.get('message');
-      let needToRefresh = false;
-      for (let i = 0; i < messagesToRefreshList.length; i++) {
-        if (message.indexOf(messagesToRefreshList[i]) > -1) {
-          needToRefresh = true;
-          break;
-        }
-      }
-
-      if (needToRefresh) {
-        this.send('refreshList');
-      }
-    }
   }
 });